@@ -1,9 +1,13 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '3.0.2'
+ruby '3.1.2'
 
 gem 'sassc-rails', '>= 2.1.0'
+
+gem 'will_paginate', '3.3.1'
+
+gem 'bootstrap-will_paginate'
 
 # Bundle edge Rails instead: gem "rails", github: "rails/rails", branch: "main"
 gem 'rails', '~> 7.0.3', '>= 7.0.3.1'
@@ -44,13 +48,16 @@ gem 'tzinfo-data', platforms: %i[mingw mswin x64_mingw jruby]
 # Reduces boot times through caching; required in config/boot.rb
 gem 'bootsnap', require: false
 
-gem 'turbo-rails'
-
 gem 'wicked_pdf'
+
+gem 'wkhtmltopdf-binary'
 
 gem 'bootstrap-sass', '~> 3.4.1'
 
 gem 'jquery-rails'
+
+gem 'bcrypt', '~> 3.1.7'
+
 
 # Use Active Storage variants [https://guides.rubyonrails.org/active_storage_overview.html#transforming-images]
 # gem "image_processing", "~> 1.2"
