@@ -1,5 +1,10 @@
 Rails.application.routes.draw do
-  root to: 'pages#index'
-  get 'about', to: 'pages#about'
   resources :articles
+  root 'pages#index'
+  get 'about', to: 'pages#about'
+  get 'signup', to: 'users#new'
+  resources :users, except: [:new]
+  get 'login', to: 'sessions#new'
+  post 'login', to: 'sessions#create'
+  delete 'logout', to: 'sessions#destroy'
 end
